+++ src/entropy/mod.rs
@@ -2,10 +2,12 @@
 
 mod collision;
 mod hartley;
+mod min_entropy;
 mod renyi;
 mod shannon;
 
 pub use collision::{collision_entropy, collision_entropy_distribution};
 pub use hartley::{hartley_entropy, hartley_entropy_distribution};
+pub use min_entropy::{min_entropy, min_entropy_distribution};
 pub use renyi::{InvalidRenyiOrder, renyi_entropy, renyi_entropy_distribution};
 pub use shannon::{shannon, shannon_distribution};